@@ -19,13 +19,12 @@
     "gray-matter": "^4.0.3",
     "mermaid": "^9.1.6",
     "minisearch": "^6.0.0",
-    "next": "12",
-    "next-plausible": "^3.1.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "next": "13.1.6",
+    "next-plausible": "^3.7.1",
+    "react": "18.2.0",
+    "react-dom": "^18.2.0",
     "remark": "^13.0.0",
     "remark-highlight.js": "^6.0.0",
-    "remark-html": "^13.0.1",
-    "swr": "^0.5.6"
+    "remark-html": "^13.0.1"
   }
 }
