@@ -1,41 +1,40 @@
-const Feed = require('feed').Feed;
-import remark from 'remark'
-import html from 'remark-html'
-const fs = require('fs');
-import { getSortedPostsDataWithHtml } from './posts'
+const Feed = require("feed").Feed;
+import remark from "remark";
+import html from "remark-html";
+const fs = require("fs");
+import { getSortedPostsDataWithHtml } from "./posts";
 
 export async function generateRssFeed(tag) {
-  if (process.env.NODE_ENV === 'development') {
+  if (process.env.NODE_ENV === "development") {
     return;
   }
-
 
   const baseUrl = "https://daniel.feldroy.com";
   const date = new Date();
   const author = {
-    name: 'Daniel Roy Greenfeld',
-    email: 'daniel@feldroy.com',
-    link: 'https://daniel.feldroy.com'
+    name: "Daniel Roy Greenfeld",
+    email: "daniel@feldroy.com",
+    link: "https://daniel.feldroy.com",
   };
 
   const feed = new Feed({
     title: `Daniel Roy Greenfeld`,
-    description: 'Inside the head of Daniel Roy Greenfeld',
-    id: baseUrl + '/',
+    description: "Inside the head of Daniel Roy Greenfeld",
+    id: baseUrl + "/",
     link: baseUrl,
-    language: 'en',
+    language: "en",
     image: `${baseUrl}/images/pydanny-cartwheel.png`,
     favicon: `${baseUrl}/favicon.ico`,
     copyright: `All rights reserved ${date.getFullYear()}, Daniel Roy Greenfeld`,
     updated: date,
-    generator: 'Next.js using Feed for Node.js',
+    generator: "Next.js using Feed for Node.js",
     feedLinks: {
-      rss2: `${baseUrl}/feeds/feed.xml`
+      rss2: `${baseUrl}/feeds/feed.xml`,
     },
-    author
+    author,
   });
 
-  const posts = getSortedPostsDataWithHtml().slice(0, 2)
+  const posts = getSortedPostsDataWithHtml().slice(0, 2);
 
   for (const post of posts) {
     if (tag.length > 0 && post.tags.indexOf(tag) === -1) {
@@ -43,10 +42,8 @@ export async function generateRssFeed(tag) {
     }
 
     const url = `${baseUrl}/posts/${post.id}`;
-    const processedContent = await remark()
-      .use(html)
-      .process(post.content)
-    const contentHtml = processedContent.toString()
+    const processedContent = await remark().use(html).process(post.content);
+    const contentHtml = processedContent.toString();
 
     feed.addItem({
       title: post.title,
@@ -56,11 +53,11 @@ export async function generateRssFeed(tag) {
       content: contentHtml,
       author: [author],
       contributor: [author],
-      date: new Date(post.date)
+      date: new Date(post.date),
     });
-  };
+  }
 
-  fs.mkdirSync('./public/feeds', { recursive: true });
+  fs.mkdirSync("./public/feeds", { recursive: true });
   if (tag.length > 0) {
     fs.writeFileSync(`./public/feeds/${tag}.atom.xml`, feed.atom1());
   } else {
